Refactor app-* modules to android-* and migrate to kts style
