+++ android-auth/src/main/AndroidManifest.xml
