+++ android-auth/src/main/res/values/strings.xml
