+++ android-onboarding/build.gradle.kts
@@ -1,3 +1,5 @@
+import co.anitrend.buildSrc.Libraries
+
 /*
  * Copyright (C) 2019  AniTrend
  *
@@ -16,17 +18,10 @@
  */
 
 plugins {
-    id("com.android.library")
-    id("kotlin-android")
-    id("kotlin-kapt")
-    id("kotlin-android-extensions")
+    id("co.anitrend.plugin")
 }
 
-apply from: "../gradle/gradle-common-config.gradle"
-apply from: "../gradle/gradle-common-config-app.gradle"
-apply from: "../gradle/gradle-common-dependencies.gradle"
-
-
 dependencies {
-
+    implementation(Libraries.liquidSwipe)
+    implementation(Libraries.AirBnB.Lottie.lottie)
 }
