+++ android-onboarding/src/main/AndroidManifest.xml
