+++ android-onboarding/src/main/kotlin/co/anitrend/onboarding/model/OnBoarding.kt
