+++ android-onboarding/src/main/kotlin/co/anitrend/onboarding/ui/fragment/OnBoardingContent.kt
@@ -22,8 +22,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.lifecycleScope
-import co.anitrend.arch.extension.argument
-import co.anitrend.arch.extension.getCompatColor
+import co.anitrend.arch.extension.ext.argument
+import co.anitrend.arch.extension.ext.getCompatColor
 import co.anitrend.arch.ui.fragment.SupportFragment
 import co.anitrend.core.koin.helper.DynamicFeatureModuleHelper
 import co.anitrend.core.ui.fragment.AniTrendFragment
