+++ android-onboarding/src/main/res/drawable/gradient_slide_01.xml
