+++ android-onboarding/src/main/res/drawable/gradient_slide_02.xml
