+++ android-onboarding/src/main/res/drawable/gradient_slide_03.xml
