+++ android-onboarding/src/main/res/drawable/gradient_slide_04.xml
