+++ android-onboarding/src/main/res/drawable/gradient_slide_05.xml
