+++ android-onboarding/src/main/res/drawable/gradient_slide_06.xml
