+++ android-onboarding/src/main/res/layout/onboarding_content.xml
