+++ android-onboarding/src/main/res/layout/onboarding_screen.xml
