+++ android-onboarding/src/main/res/raw/bookmark.json
