+++ android-onboarding/src/main/res/raw/cartoon_loading.json
