+++ android-onboarding/src/main/res/raw/open_book.json
