+++ android-onboarding/src/main/res/raw/rating_system.json
