+++ android-onboarding/src/main/res/raw/swipe_left.json
