+++ android-onboarding/src/main/res/raw/writing_message.json
