+++ android-onboarding/src/main/res/values/strings.xml
