+++ android-onboarding/src/main/res/values/theme.xml
