+++ android-splash/.gitignore
