+++ android-splash/src/main/AndroidManifest.xml
