+++ android-splash/src/main/kotlin/co/anitrend/splash/presenter/SplashPresenter.kt
