+++ android-splash/src/main/kotlin/co/anitrend/splash/ui/fragment/SplashContent.kt
