+++ android-splash/src/main/res/drawable/stripe_decoration_rounded.xml
