+++ android-splash/src/main/res/layout/activity_splash.xml
