+++ android-splash/src/main/res/layout/content_splash.xml
@@ -103,14 +103,14 @@
         app:layout_constraintTop_toBottomOf="@+id/appCompatImageView"
         app:layout_constraintVertical_bias="0.0">
 
-        <co.anitrend.arch.ui.view.text.SingleLineTextView
+        <com.google.android.material.textview.MaterialTextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/splash_label_segment_first"
             android:textAppearance="@style/TextAppearance.MaterialComponents.Headline5"
             android:textStyle="bold" />
 
-        <co.anitrend.arch.ui.view.text.SingleLineTextView
+        <com.google.android.material.textview.MaterialTextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/splash_label_segment_second"
@@ -130,7 +130,7 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent">
 
-        <co.anitrend.arch.ui.view.text.SingleLineTextView
+        <com.google.android.material.textview.MaterialTextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
@@ -138,7 +138,7 @@
             android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
             android:textStyle="bold" />
 
-        <co.anitrend.arch.ui.view.text.SingleLineTextView
+        <com.google.android.material.textview.MaterialTextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
