+++ android-splash/src/main/res/values/dimens.xml
