+++ android-splash/src/main/res/values/strings.xml
