--- app-auth/build.gradle
@@ -1,59 +1,0 @@
-/*
- * Copyright (C) 2019  AniTrend
- *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     GNU General Public License for more details.
- *
- *     You should have received a copy of the GNU General Public License
- *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-plugins {
-    id("com.android.library")
-    id("kotlin-android")
-    id("kotlin-kapt")
-    id("kotlin-android-extensions")
-}
-
-apply from: "../gradle/gradle-common-config.gradle"
-apply from: "../gradle/gradle-common-dependencies.gradle"
-
-android {
-    defaultConfig {
-        defaultConfig {
-            consumerProguardFiles 'consumer-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    /** Project Modules */
-    implementation project(':app-core')
-    implementation project(':app-data')
-    implementation project(':app-domain')
-    implementation project(':android-core')
-
-    /** Platform Dependencies */
-    implementation project(":support-domain")
-    implementation project(":support-theme")
-    implementation project(":support-core")
-    implementation project(":support-ext")
-    implementation project(":support-ui")
-
-    /** Koin AndroidX Dependency Injection */
-    implementation("org.koin:koin-androidx-viewmodel:$koin")
-
-    /** Material Design */
-    implementation("com.google.android.material:material:$material")
-
-    /** Testing-only dependencies */
-    testImplementation("org.koin:koin-test:$koin")
-    androidTestImplementation("androidx.test.espresso:espresso-core:$espresso")
-}
